+++ src/App.js
@@ -4,6 +4,7 @@ import Home from "Components/Home/Home";
 import Login from "Components/Login/Login";
 import Signup from "Components/Signup/Signup";
 import PrivateRoute from "hoc/PrivateRoute";
+import Error from "Components/Error/Error";
 
 export default function App() {
   return (
@@ -14,19 +15,7 @@ export default function App() {
         <Route path="/home" element={<PrivateRoute />}>
           <Route path="/home" element={<Home />} />
         </Route>
-        <Route
-          path="*"
-          element={
-            <div>
-              <p style={{ fontSize: "35px", fontWeight: "900" }}>
-                404 <br /> Page not found
-              </p>
-              <a style={{ fontSize: "20px" }} href="/">
-                Go to home
-              </a>
-            </div>
-          }
-        />
+        <Route path="*" element={<Error />} />
       </Routes>
     </div>
   );
