+++ src/Components/Home/Home.scss
@@ -1,29 +1,5 @@
 .home {
   display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  &__btn {
-    font-size: 1.4rem;
-    font-weight: 600;
-    width: 32rem;
-    height: 4rem;
-    background-color: #fdf5f2;
-    border-radius: 3px;
-    margin-top: 5rem;
-    cursor: pointer;
-    color: #eb7575;
-    border: none;
-    box-shadow: rgb(15 15 15 / 10%) 0px 1px 2px,
-      rgb(235 87 87 / 30%) 0px 0px 0px 1px inset;
-    transition: all 0.3s;
-
-    &:hover {
-      background-color: #f8f8f8;
-    }
-  }
-
-  &__img {
-    width: 30rem;
-  }
+  width: 100vw;
+  height: 100vh;
 }
