+++ src/Components/Sidebar/Sidebar.scss
@@ -1,0 +1,34 @@
+.sidebar {
+  background-color: #202020;
+  width: 17%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+
+  &__logo {
+    width: 15rem;
+    height: 10rem;
+
+    display: flex;
+    align-self: start;
+  }
+
+  &__btn {
+    font-size: 1.4rem;
+    font-weight: 600;
+    height: 4rem;
+    background-color: #fdf5f2;
+    border-radius: 3px;
+    margin: 1rem;
+    cursor: pointer;
+    color: #eb7575;
+    border: none;
+    box-shadow: rgb(15 15 15 / 10%) 0px 1px 2px,
+      rgb(235 87 87 / 30%) 0px 0px 0px 1px inset;
+    transition: all 0.3s;
+
+    &:hover {
+      background-color: #f8f8f8;
+    }
+  }
+}
