+++ src/Components/SlateEditor/SlateEditor.scss
@@ -6,7 +6,7 @@
     padding: 3rem;
     font-size: 1.7rem;
     text-align: left;
-    background-color: #353230;
+    background-color: #191919;
     color: #f2f2f2;
     min-height: 100vh;
   }
